cli: move config-reading to new `config` module
