+++ src/lib.rs
@@ -18,6 +18,7 @@
 extern crate pest_derive;
 
 pub mod commands;
+pub mod config;
 pub mod diff_edit;
 pub mod formatter;
 pub mod graphlog;
